(wip) start condensing ib clustering
