+++ hydra/src/openset/CMakeLists.txt
@@ -3,4 +3,6 @@ target_sources(
   PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/embedding_distances.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/embedding_group.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/embedding_pooling.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/clustering/agglomerative_clustering.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/clustering/ib_utilities.cpp
 )
